policies + route model binding + syntax fixes

+++ src/Http/Controllers/SettingController.php
@@ -2,22 +2,29 @@
 
 namespace PictaStudio\Contento\Http\Controllers;
 
+use Illuminate\Http\Resources\Json\{AnonymousResourceCollection, JsonResource};
+use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Gate;
 use PictaStudio\Contento\Http\Requests\SaveSettingRequest;
 use PictaStudio\Contento\Http\Resources\SettingResource;
 use PictaStudio\Contento\Models\Setting;
 
 class SettingController extends Controller
 {
-    public function index()
+    public function index(): AnonymousResourceCollection
     {
+        Gate::authorize('viewAny', Setting::class);
+
         $settings = Setting::all();
 
         return SettingResource::collection($settings);
     }
 
-    public function store(SaveSettingRequest $request)
+    public function store(SaveSettingRequest $request): JsonResource
     {
+        Gate::authorize('create', Setting::class);
+
         $setting = Setting::updateOrCreate(
             [
                 'group' => $request->group,
@@ -29,9 +36,10 @@ class SettingController extends Controller
         return new SettingResource($setting);
     }
 
-    public function destroy($id)
+    public function destroy(Setting $setting): Response
     {
-        $setting = Setting::findOrFail($id);
+        Gate::authorize('delete', $setting);
+
         $setting->delete();
 
         return response()->noContent();
